web/requirements.txt
flask
deta
python-dotenv
werkzeug
